Restored ApiTest base class functionality
